Controlando os dados por usuário logado

+++ curso-jsp/src/main/java/dao/DAOUsuarioRepository.java
@@ -20,18 +20,19 @@ public class DAOUsuarioRepository {
 	}
 	
 	/*faz o inser no banco da página de cadastro usuario.jsp*/
-	public ModelLogin gravarUser(ModelLogin objeto) throws Exception {
+	public ModelLogin gravarUser(ModelLogin objeto, Long userLogado) throws Exception {
 		
 		if(objeto.isNovo()) {/*Grava um novo*/
 			
 		
-			String sql = "INSERT INTO model_login(login, senha, nome, email) VALUES (?, ?, ?, ?)";
+			String sql = "INSERT INTO model_login(login, senha, nome, email, usuario_id) VALUES (?, ?, ?, ?, ?)";
 	
 			PreparedStatement insert = connection.prepareStatement(sql);
 			insert.setString(1, objeto.getLogin());
 			insert.setString(2, objeto.getSenha());
 			insert.setString(3, objeto.getNome());
 			insert.setString(4, objeto.getEmail());
+			insert.setLong(5, userLogado);
 			insert.execute();
 			connection.commit();
 			
@@ -47,15 +48,15 @@ public class DAOUsuarioRepository {
 			connection.commit();
 			
 		}
-			return this.consultaUsuario(objeto.getLogin());
+			return this.consultaUsuario(objeto.getLogin(), userLogado);
 		
 	}
 	
-	public List<ModelLogin> consultaUsuarioList() throws Exception{
+	public List<ModelLogin> consultaUsuarioList(Long userLogado) throws Exception{
 		
 		List<ModelLogin> retorno = new ArrayList<ModelLogin>();
 		
-		String sql = "SELECT * FROM model_login";
+		String sql = "SELECT * FROM model_login where useradmin is false and usuario_id = " + userLogado;
 		PreparedStatement statement = connection.prepareStatement(sql);
 		
 		ResultSet resultSet = statement.executeQuery();
@@ -75,13 +76,14 @@ public class DAOUsuarioRepository {
 		
 	}
 	
-	public List<ModelLogin> consultaUsuarioList(String nome) throws Exception{
+	public List<ModelLogin> consultaUsuarioList(String nome, Long userLogado) throws Exception{
 		
 		List<ModelLogin> retorno = new ArrayList<ModelLogin>();
 		
-		String sql = "SELECT * FROM model_login where upper(nome) like upper(?)";
+		String sql = "SELECT * FROM model_login where upper(nome) like upper(?) and useradmin is false and usuario_id = ?";
 		PreparedStatement statement = connection.prepareStatement(sql);
 		statement.setString(1, "%" + nome + "%");
+		statement.setLong(2, userLogado);
 		
 		ResultSet resultSet = statement.executeQuery();
 		
@@ -101,7 +103,29 @@ public class DAOUsuarioRepository {
 		
 	}
 	
-	public ModelLogin consultaUsuario(String login) throws Exception {
+	public ModelLogin consultaUsuario(String login, Long userLogado) throws Exception {
+		
+		ModelLogin modelLogin = new ModelLogin();
+		
+		String sql = "select * from model_login where upper(login) = upper('"+ login +"') and useradmin is false and usuario_id = " + userLogado;
+		PreparedStatement statement = connection.prepareStatement(sql);
+		
+		ResultSet resultSet = statement.executeQuery();
+		
+		while (resultSet.next()) /*se tem resultado*/ {
+			
+			modelLogin.setId(resultSet.getLong("id"));
+			modelLogin.setEmail(resultSet.getString("email"));
+			modelLogin.setLogin(resultSet.getString("login"));
+			modelLogin.setSenha(resultSet.getString("senha"));
+			modelLogin.setNome(resultSet.getString("nome"));
+		}
+		
+		return modelLogin;
+		
+	}
+	
+	public ModelLogin consultaUsuarioLogado(String login) throws Exception {
 		
 		ModelLogin modelLogin = new ModelLogin();
 		
@@ -123,13 +147,36 @@ public class DAOUsuarioRepository {
 		
 	}
 	
-		public ModelLogin consultaUsuarioID(String id) throws Exception {
+	public ModelLogin consultaUsuario(String login) throws Exception {
+		
+		ModelLogin modelLogin = new ModelLogin();
+		
+		String sql = "select * from model_login where upper(login) = upper('"+ login +"') and useradmin is false";
+		PreparedStatement statement = connection.prepareStatement(sql);
+		
+		ResultSet resultSet = statement.executeQuery();
+		
+		while (resultSet.next()) /*se tem resultado*/ {
+			
+			modelLogin.setId(resultSet.getLong("id"));
+			modelLogin.setEmail(resultSet.getString("email"));
+			modelLogin.setLogin(resultSet.getString("login"));
+			modelLogin.setSenha(resultSet.getString("senha"));
+			modelLogin.setNome(resultSet.getString("nome"));
+		}
+		
+		return modelLogin;
+		
+	}
+	
+		public ModelLogin consultaUsuarioID(String id, Long userLogado) throws Exception {
 				
 				ModelLogin modelLogin = new ModelLogin();
 				
-				String sql = "select * from model_login where id = ?";
+				String sql = "select * from model_login where id = ? and useradmin is false and usuario_id = ?";
 				PreparedStatement statement = connection.prepareStatement(sql);
 				statement.setLong(1, Long.parseLong(id));
+				statement.setLong(2, userLogado);
 				
 				ResultSet resultSet = statement.executeQuery();
 				
@@ -166,7 +213,7 @@ public class DAOUsuarioRepository {
 	
 	public void deletarUser(String idUser) throws Exception {
 		
-		String sql = "DELETE FROM model_login WHERE id = ? ";
+		String sql = "DELETE FROM model_login WHERE id = ? and useradmin is false";
 		PreparedStatement deletar = connection.prepareStatement(sql);
 		deletar.setLong(1, Long.parseLong(idUser));
